third_party/blink/web_tests/fast/scroll-behavior/bordered-container-child-scroll-expected.txt

Issue crbug.com/295848: element.scrollIntoViewIfNeeded() does not scroll the entire element render box into view.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS parent.scrollTop is 50
PASS parent.scrollLeft is 50
PASS parent.scrollTop is 0
PASS parent.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE

	
	
		
